Add a helper to format and display packages informations and avoid forwarding to the info command
